Split out misc routines per Max
Previously I thought that the Floating Point stuff was "boxed into" the
Misc Int Routines by the presence of kcReturnFromException and
LoadInterruptRegisters in v02.28. I now suspect that they are actually
part of the Int Handlers in v02.28, and that there is therefore no
reason to group the Floating Point stuff with the Misc stuff.